v11_multimodal_embedding/requirements.txt
torch>=2.1.0
numpy>=1.24.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
